# Faster tokenization (log(n) lookups & caching of possible merges)