This greatly speeds up tokenization of large prompts (up to 3000x faster for 2k prompt tokens), making it possible to use prompts of the full context length without paying quadratic costs.

Tested `runfast` builds on an M2 Max Mac w/ [hyperfine](https://github.com/sharkdp/hyperfine) & [lorem](https://formulae.brew.sh/formula/lorem):
```bash
# `lorem -n 128` yields 336 tokens;
# Only generating / consuming one token to test tokenization itself
hyperfine -L bin run_fasttoken,run_master --warmup=1 "./{bin} llama2.c.stories110M.bin -n 1 -i \"\$( lorem -n 128 )\""
```

Result:
```
Benchmark 1: ./run_fasttoken "llama2.c.stories110M.bin" -n 1 -i "$( lorem -n 128 )"
  Time (mean ± σ):      86.7 ms ±   2.9 ms    [User: 32.5 ms, System: 31.0 ms]
  Range (min … max):    81.8 ms …  90.6 ms    10 runs
 
Benchmark 2: ./run_master "llama2.c.stories110M.bin" -n 1 -i "$( lorem -n 128 )"
  Time (mean ± σ):     11.151 s ±  0.177 s    [User: 10.927 s, System: 0.168 s]
  Range (min … max):   10.950 s … 11.563 s    10 runs
 
Summary
  './run_fasttoken "llama2.c.stories110M.bin" -n 1 -i "$( lorem -n 128 )"' ran
  128.61 ± 4.72 times faster than './run_master "llama2.c.stories110M.bin" -n 1 -i "$( lorem -n 128 )"'
```

With `lorem -n 750` (1997 tokens), this PR's build takes 100ms and master takes ~6min.