@ochafik 2023-08-08
Just realized there's a simpler / even faster way to find the potential merges as any merged token will appear just after the first token of a merged pair in the sorted list of tokens. Turned this PR to Draft mode while I update the code accordingly :-)