@atamurad 2023-08-12
Thank you for optimizations! 

I've also implemented sorted vocabulary for faster lookup but on top of utf8 branch/PR. 

I used `bsearch()`, it is part of the standard C library, and we can re-use  `compare_token` for `qsort`.

https://github.com/atamurad/llama2.c/commit/8a179f69dae5b8501ce5947681b27d16e4b31db4