@karpathy 2023-08-16
We merged an alternative that incorporated some of this code. Ty!