@ochafik 2023-08-17
@karpathy @atamurad glad that helped, thanks!!